Windows setup

+++ test/helper.hpp
@@ -1,5 +1,5 @@
 #include "catch.hpp"
 #include "vore.hpp"
 
-void IS_MATCH(Match m, u_int64_t offset, u_int64_t len, std::string val);
-void SINGLE_MATCH(std::vector<MatchGroup> results, u_int64_t fileOffset, u_int64_t match_length, std::string value);
+void IS_MATCH(Match m, long long offset, long long len, std::string val);
+void SINGLE_MATCH(std::vector<MatchGroup> results, long long fileOffset, long long match_length, std::string value);
